+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 
 - Renamed `<Sparkbar />` to `<SparkBarChart />`.
 - Renamed `<Sparkline />` to `<SparkLineChart />`.
+- Removed `area`, `lineStyle`, and `hasPoint` from `SparkLineChartProps.series`.
+- Moved `offsetLeft` & `offsetRight` from `SparkLineChartProps.series` to `SparkLineChartProps`.
+- Renamed `series` to `data` in `SparkLineChartProps`.
 
 ## [0.25.0] - 2021-11-12